package structure refactor
