+++ computationEngine/src/main/java/uk/ac/cam/cl/juliet/computationengine/utility/ComplexVector.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.juliet.computationengine.range;
+package uk.ac.cam.cl.juliet.computationengine.utility;
 
 import java.util.ArrayList;
 import java.util.List;
